Rename FlMouseCursorPlugin to FlMouseCursorHandler
This matches the naming used in the other embedders and removes
the confusion between this and Flutter plugins.
