+++ shell/platform/linux/BUILD.gn
@@ -119,7 +119,7 @@ source_set("flutter_linux_sources") {
     "fl_method_channel.cc",
     "fl_method_codec.cc",
     "fl_method_response.cc",
-    "fl_mouse_cursor_plugin.cc",
+    "fl_mouse_cursor_handler.cc",
     "fl_pixel_buffer_texture.cc",
     "fl_platform_plugin.cc",
     "fl_plugin_registrar.cc",
